+++ CHANGELOG.md
@@ -7,6 +7,9 @@ Version 1.0.9
 * Add parsing of independent padding attributes
 * pstsTextAlpha value range change from 0..1 to 0..255
 * New attr pstsTextColorSelected
+* Resource.Id.tab_title changed to Resource.Id.psts_tab_title
+* Resource.Layout.tab changed to Resource.Layout.psts_tab
+* Resource.Drawable.background_tab changed to Resource.Drawable.psts_background_tab
 
 
 Version 1.0.1